Started to add tests
