--- controlfiles/artscomponents/geodetic/TestBasicGeodeticAccuracy.arts
@@ -1,19 +1,0 @@
-#DEFINITIONS:  -*-sh-*-
-#
-# Runs TestBasicGeodeticAccuracy
-
-# Patrick Eriksson 2022-09-25
-
-Arts2 {
-    # Test with WGS84
-    refellipsoidEarthZZZ(model="WGS84")
-
-    NumericCreate(max_dl)
-    VectorCreate(max_dpos)
-    VectorCreate(max_dlos)
-
-    TestBasicGeodeticAccuracy(rte_pos, max_dl, max_dpos, max_dlos, refellipsoidZZZ, 200000, 0.1)
-    Print(rte_pos)
-    Print(max_dpos)
-    Print(max_dlos)
-}
