+++ tests/core/geodetic/TestIntersections.arts
@@ -6,6 +6,7 @@
 # The methods (except the surface one) have internal asserts checking
 # the results so the script performs also some testing
 #
+
 Arts2 {
