+++ tests/core/ppath/TestPpathAddGridCrossings.arts
@@ -1,0 +1,42 @@
+#DEFINITIONS:  -*-sh-*-
+#
+#   Checks that ppathAddGridCrossings runs as expected for a dummy case
+#
+Arts2 {
+
+# WGS84 Earth and constant 0m surface elevation
+refellipsoidEarthZZZ(model = "WGS84")
+Copy(refellipsoid, refellipsoidZZZ)  # Temporary fix
+LatLonFieldSetConstant(gfield2 = surface_elevation,
+                       value = 0)
+
+# Assumed TOA altitude
+NumericCreate(z_toa)
+NumericSet(z_toa, 9.9e3)
+
+# Sensor inside the atmosphere and looking upwards
+VectorSet(rte_pos, [6, 12, 67])
+VectorSet(rte_los, [88, 80])
+#
+ppathGeometric(z_toa = z_toa, 
+               ppath_lstep = 200e3)
+
+VectorCreate(z_grid)
+VectorSet(z_grid, [5e3, 7.5e3, 8e3])  # Shall add 3 points
+
+VectorCreate(lat_grid)
+VectorSet(lat_grid, [12])  # Not passed, no extra point
+
+VectorCreate(lon_grid)
+VectorSet(lon_grid, [68, 68.01]) # Shall add 2 points
+
+ppathAddGridCrossings(ppath_lstep = 250e3,
+                      z_grid = z_grid,
+                      lat_grid = lat_grid,
+                      lon_grid = lon_grid)
+# Print( ppath, 0 )
+
+# End points laways included so np shall end up as 7
+ppathCheckEndPoint(background = "Space",
+                   np = 7)
+}
